+++ ext/SatelliteAnalysisDecayExt/main.jl
@@ -184,6 +184,12 @@ function _decay_analysis(
     )
 
     metadata!(df, "Description", "Mean orbital element evolution during the orbital decay.")
+
+    # The style `:note` makes the metadata propagate through DataFrame transformations.
+    metadata!(df, "Satellite Mass",      satellite_mass;      style = :note)
+    metadata!(df, "Satellite Mean Area", satellite_mean_area; style = :note)
+    metadata!(df, "Terminate Altitude",  terminate_altitude;  style = :note)
+
     colmetadata!(df, :date,             "Unit", :UTC)
     colmetadata!(df, :time,             "Unit", time_unit)
     colmetadata!(df, :f107,             "Unit", :sfu)