+++ ext/SatelliteAnalysisMakieExt/SatelliteAnalysisMakieExt.jl
@@ -1,9 +1,10 @@
 module SatelliteAnalysisMakieExt
 
+using Dates
 using Makie
 using Makie: Colorant
 using SatelliteAnalysis
-import SatelliteAnalysis: makie_theme, makie_palette
+import SatelliteAnalysis: makie_theme, makie_palette, plot_decay_analysis
 
 ############################################################################################
 #                                        Constants                                         #
@@ -15,6 +16,7 @@ include("./constants.jl")
 #                                         Includes                                         #
 ############################################################################################
 
+include("./decay_analysis.jl")
 include("./theme.jl")
 
 end # module SatelliteAnalysisMakieExt